requirements.txt
# Core dependencies
pandas>=2.0.0
numpy==1.24.0
transformers==4.34.0
torch==2.0.1
sentence-transformers>=2.2.2
chromadb>=0.4.18
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
accelerate

# Utilities
tqdm>=4.66.0
typing-extensions>=4.7.0
pydantic>=2.4.0

# Logging
rich>=13.5.0
